Create a method Sphere.navigate(x, y) that tells it to go somewhere
When a new Sphere object is instantiated, the Sphere.create constructor now also expects a reference to the current Scene object.  Sphere objects now also have a public method Sphere.updateAi which should be called from outside and will recompute the path the goal if a goal has been defined.  I decided that from now on every object in the scene should have an updateAi method.