Create a class ProjectileRenderer
A ProjectileRenderer can render projectiles but does not itself cache any game state related to the projectiles. Previously I put the code for both the rendering of an object and its behaviour into the same class which turned out to be a huge mistake, so now I'm trying to learn from this mistake by having separate classes Projectile and ProjectileRenderer.